+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Next
+
+- **[Feature]** Add `join` function.
+- **[Feature]** Add `asWritableUrl` function.
+- **[Feature]** Add `asFuri` function.
+
 # 1.1.0 (2018-12-03)
 
 - **[Feature]** Publish `.mjs` files.